Fixed bug with balance type in db

+++ dbtypes.go
@@ -2,7 +2,7 @@ package main
 
 type User struct {
 	//	ID         primitive.ObjectID `bson:"_id" json:"_id"`
-	Username   string `bson:"username" json:"username"`
-	TelegramID int    `bson:"telegramID" json:"telegramID"`
-	Balance    int    `bson:"balance" json:"balance"`
+	Username   string  `bson:"username" json:"username"`
+	TelegramID int     `bson:"telegramID" json:"telegramID"`
+	Balance    float64 `bson:"balance" json:"balance"`
 }
